requirements.txt
Appium-Client-Python
selenium